+++ src/Adapter/Factory/SftpAdapterFactory.php
@@ -3,19 +3,19 @@
 namespace BsbFlysystem\Adapter\Factory;
 
 use BsbFlysystem\Exception\RequirementsException;
+use BsbFlysystem\Exception\UnexpectedValueException;
 use League\Flysystem\Sftp\SftpAdapter as Adapter;
 use Zend\ServiceManager\FactoryInterface;
 use Zend\ServiceManager\ServiceLocatorInterface;
-use UnexpectedValueException;
 
-class SftpAdapterFactory extends AbstractAdapterFactory implements FactoryInterface
+class SftpAdapterFactory extends AbstractAdapterFactory
 {
     /**
      * @inheritdoc
      */
     public function doCreateService(ServiceLocatorInterface $serviceLocator)
     {
-        if (!class_exists('League\Flysystem\Sftp\SftpAdapter')) {
+        if (!class_exists(\League\Flysystem\Sftp\SftpAdapter::class)) {
             throw new RequirementsException(
                 ['league/flysystem-sftp'],
                 'Sftp'