+++ src/Exception/RuntimeException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace BsbFlysystem\Exception;
+
+class RuntimeException extends \RuntimeException
+{
+}